refactor : useApiClient를 사용하지 않고 useSuspenseQuery를 이용해 급상승 랭킹 아이템을 구현 - useSuspenseQuery를 이용한 이유 : data에 값이 존재하는 것이 보장되기 때문에 Optional Chaining할 필요가 없고, Suspense를 활용한 isLoading 처리를 간편하게 함
