+++ src/components/Home/PresentRanking/Cotainer/useRankingItem.ts
@@ -1,7 +1,7 @@
 import { getBasicFetch } from '@src/api/getBasicFetch';
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
-import type { Good } from '@src/types/Goods';
-import { useQuery } from '@tanstack/react-query';
+import type { Goods } from '@src/types/Goods';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
 export const useRankingItem = ({
   targetType,
@@ -10,14 +10,13 @@ export const useRankingItem = ({
   targetType: string | null;
   rankType: string | null;
 }) => {
-  const { data, isError, isLoading } = useQuery<Good[]>({
+  const { data, isError, isLoading } = useSuspenseQuery<Goods>({
     queryKey: ['ranking', { targetType, rankType }],
-    queryFn: () => getBasicFetch<Good[]>(BASIC_ENDPOINT.ranking),
-    params: {
-      targetType,
-      rankType,
-    },
-    enabled: !!targetType && !!rankType,
+    queryFn: () =>
+      getBasicFetch<Goods>(BASIC_ENDPOINT.ranking, {
+        targetType,
+        rankType,
+      }),
   });
   return { data, isError, isLoading };
 };
